+++ ads/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
-from ads.models import Location, Users, Categories, Ad, Selection
-
+from ads.models import Location, Categories, Ad, Selection
+from authentification.models import User
 
 class LocationSerializers(serializers.ModelSerializer):
     class Meta:
@@ -13,62 +13,8 @@ class CategoriesSerializers(serializers.ModelSerializer):
         model = Categories
         fields = '__all__'
 
-class UsersSerializers(serializers.ModelSerializer):
-    location = serializers.SlugRelatedField(many=True, read_only=True, slug_field="name")
-    class Meta:
-        model = Users
-        fields = ["id", "first_name", "last_name", "username", "password", "role", "age", "age", "location"]
-
-class UsersCreateSerializers(serializers.ModelSerializer):
-    id = serializers.IntegerField(required=False)
-    location = serializers.SlugRelatedField(many=True,
-                                          queryset = Location.objects.all(),
-                                          slug_field="name",
-                                          required=False)
-    class Meta:
-        model = Users
-        fields = ["id", "first_name", "last_name", "username", "password", "role", "age", "age", "location"]
-
-    def is_valid(self, *, raise_exception=False):
-        self._location = self.initial_data.pop("location")
-        return super().is_valid(raise_exception=raise_exception)
-
-    def create(self, validated_data):
-        user = Users.objects.create(**validated_data)
-        for location in self._location:
-            location_obj, _ = Location.objects.get_or_create(name=location)
-            user.location.add(location_obj)
-        user.save()
-        return user
-
-class UsersUpdateSerializers(serializers.ModelSerializer):
-    location = serializers.SlugRelatedField(many=True,
-                                          queryset = Location.objects.all(),
-                                          slug_field="name",
-                                          required=False)
-    class Meta:
-        model = Users
-        fields = ["id", "first_name", "last_name", "username", "password", "role", "age", "age", "location_id"]
-
-    def is_valid(self, *, raise_exception=False):
-        self._location = self.initial_data.pop("location")
-        return super().is_valid(raise_exception=raise_exception)
-
-    def save(self):
-        user = super().save()
-        for location in self._location:
-            location_obj, _ = Location.objects.get_or_create(name=location)
-            user.location.add(location_obj)
-        user.save()
-        return user
-
-class UsersDestroySerializers(serializers.ModelSerializer):
-    class Meta:
-        model = Users
-        fields = ["id"]
-
 class AdSerializers(serializers.ModelSerializer):
-    author = serializers.SlugRelatedField(queryset=Users.objects.all(),   slug_field="first_name",  required=False)
+    author = serializers.SlugRelatedField(queryset=User.objects.all(),   slug_field="first_name",  required=False)
     category = serializers.SlugRelatedField(queryset=Categories.objects.all(), slug_field="name", required=False)
     class Meta:
         model = Ad
@@ -104,7 +50,7 @@ class SelectionSerializers(serializers.ModelSerializer):
         fields = ["id", "name"]
 
 class SelectionAdSerializers(serializers.ModelSerializer):
-    author = serializers.SlugRelatedField(queryset=Users.objects.all(),   slug_field="id",  required=False)
+    author = serializers.SlugRelatedField(queryset=User.objects.all(),   slug_field="id",  required=False)
     category = serializers.SlugRelatedField(queryset=Categories.objects.all(), slug_field="id", required=False)
     class Meta:
         model = Ad
