+++ ads/urls.py
@@ -3,13 +3,6 @@ from ads import views
 
 urlpatterns = [
     path('', views.index),
-    path('users/', views.UsersListView.as_view()),
-    path('users/<int:pk>/', views.UsersDetailView.as_view()),
-    path('users/create/', views.UsersCreateView.as_view()),
-    path('users/<int:pk>/update/', views.UsersUpdateView.as_view()),
-    path('users/<int:pk>/delete/', views.UsersDeleteView.as_view()),
-    path('users/Z/', views.UsersZView.as_view()),
-
 
     path('ad/', views.AdListView.as_view()),
     path('ad/<int:pk>/update/', views.AdUpdateView.as_view()),
