+++ authentification/models.py
@@ -1,9 +1,21 @@
 from django.db import models
 from django.contrib.auth.models import AbstractUser
+from ads.models import Location
 
 class User(AbstractUser):
-    MALE = "m"
-    FEMALE = "f"
-    SEX = [(MALE, "Male"), (FEMALE, "Female")]
 
-    sex = models.CharField(max_length=1, choices=SEX, default=MALE)
+
+    first_name = models.CharField(max_length=50)
+    last_name = models.CharField(max_length=50)
+    role = models.CharField(max_length=50)
+    age = models.IntegerField()
+    location = models.ManyToManyField(Location)
+
+    def __str__(self):
+        return self.first_name
+
+    class Meta:
+        verbose_name = 'Пользователь'
+        verbose_name_plural = 'Пользователи'
+        ordering = ["username"]
+
